pyproject.toml
[project]
name = "booklogger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "google-auth>=2.41.1",
    "google-cloud>=0.34.0",
    "google-cloud-vision>=3.10.2",
    "gspread>=6.2.1",
    "gspread-dataframe>=4.0.0",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "zxing-cpp>=2.3.0",
]
